requirements.txt
Keras==2.1.4
numpy==1.15.4
scipy==1.1.0
scikit_image==0.14.0
matplotlib==2.2.3
tensorflow_gpu==2.0.0
Pillow==5.4.1
PyQt5==5.11.3
sip==4.19.8
tensorflow==1.12.0
opencv==4.1.2